更新页面bug1

+++ src/main/java/com/employee/system/entity/Department.java
@@ -18,6 +18,5 @@ public class Department {
     @NotNull
     private String departmentName;
 
-    private
-    Integer total;
+    private Long total;
 }
